First Upload
This is the first version of the inventory system. I tried to implement
a raycast system so that when a component is clicked on, it will be put
into the player's inventory. However, I couldn't wrap my head around 2D
raycasts. I intend to revisit this, but for now here is the early
version of the inventory.