Visible Inventory added.
We now have a visible inventory for the game. Sprites are a little on
the small side but I think that can be easily fixed. There is a test exe
in there that was made to test the size of the boxes. It can be deleted,
however I didn't want to touch any files as I'm kinda tired and I didnt
want to risk screwing things up.